+++ app/src/main/java/org/simple/clinic/home/overdue/SortedOverdueAppointment.kt
@@ -1,0 +1,7 @@
+package org.simple.clinic.home.overdue
+
+data class SortedOverdueAppointment(
+    val appointment: OverdueAppointment,
+    val score: Float,
+    val bucket: OverdueBucket
+)